Move serde-snapshot dependent accounts-db tests
